rename toggler -> notification_toggler
