+++ app/lib/components/notification_toggler.svelte
